requirements.txt
# Backend - FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25

# Validación y configuración
pydantic==2.5.3
pydantic-settings==2.1.0

# Seguridad
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Datos financieros
yfinance==1.0
pandas==2.1.4
numpy==1.26.3
ta==0.11.0

# Machine Learning - Core
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0
statsmodels==0.14.1

# Deep Learning (opcional para LSTM)
torch>=2.0.0
transformers>=4.36.0

# NLP y Sentiment Analysis
nltk==3.8.1
textblob==0.18.0

# Dashboard
streamlit==1.30.0
requests==2.31.0
plotly==5.18.0

# Utilidades
python-dotenv==1.0.0
joblib==1.3.2
email-validator==2.1.0
